move LI declaration to constants file and change file type to tsx
